require auth for conversation routes
